Alex branch

+++ src/main/java/de/learny/controller/TestController.java
@@ -8,9 +8,13 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RestController;
 
+import de.learny.controller.exception.NotEnoughPermissionsException;
 import de.learny.controller.exception.ResourceNotFoundException;
 import de.learny.dataaccess.TestRepository;
+import de.learny.domain.Account;
+import de.learny.domain.Subject;
 import de.learny.domain.Test;
+import de.learny.security.service.LoggedInAccountService;
 
 @RestController
 @RequestMapping("/api/tests")
@@ -18,6 +22,9 @@ public class TestController {
 	
 	@Autowired
 	private TestRepository testRepository;
+	
+	@Autowired
+	private LoggedInAccountService userToAccountService;
 	
 	@RequestMapping(value = "", method = RequestMethod.GET)
 	Iterable<Test> getAllTests(){
@@ -42,8 +49,12 @@ public class TestController {
 	
 	@RequestMapping(value = "/{id}", method = RequestMethod.DELETE)
 	void delete(@PathVariable("id") long id){
-		//TODO: Kann nur der Verantwortlicher
-		this.testRepository.delete(id);
+		Test test = testRepository.findById(id);
+		if(test == null)
+			throw new ResourceNotFoundException("Ein Test mit diese id existiert nicht");
+		if(permitted(id)){
+			this.testRepository.delete(id);
+		}
 	}
 	
 	@RequestMapping(value = "/{id}", method=RequestMethod.PUT, consumes={MediaType.APPLICATION_JSON_VALUE})
@@ -69,5 +80,20 @@ public class TestController {
 	@RequestMapping(value = "/{id}/highscore", method = RequestMethod.GET)
 	void getHighscoreFromTest(@PathVariable("id") long id){
 		//TODO: Muss noch implemtiert werden
+	}
+	
+	private boolean permitted(long id){
+		// Übeprüft ob Subject vorhaden
+		Test test = testRepository.findById(id);
+		// Überprüft ob Account Admin oder verantwortlich für Subject
+		Account loggedInAccount = userToAccountService.getLoggedInAccount();
+		boolean inCharge = false;
+		Subject subject = test.getSubject();
+		inCharge = subject.getAccountsInCharge().contains(loggedInAccount);
+		if (inCharge || loggedInAccount.hasRole("admin")) {
+			return true;
+		} else {
+			throw new NotEnoughPermissionsException("Nicht genug Rechte, um das auszuführen.");
+		}
 	}
 }
